+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MOS65XX arch support
 - SH arch support
 - Alpha arch support
+- HPPA arch support
 - Arch-specific features flags to enable/disable arch support
 - Expose X86 instruction encoding info via `X86InsnDetail::encoding()`
 - Make RegAccessType available for ARM64